refactor: 시간표 공유 관련 함수 한 파일로 통합 및 이름 변경
