+++ src/hooks/useTableShare.tsx
@@ -2,7 +2,7 @@ import { useModal } from './useModal';
 import ShareModal from '../components/ShareModal/ShareModal';
 import { useGetDebateTableData } from './query/useGetDebateTableData';
 import { useEffect, useState } from 'react';
-import { createTableShareUrl } from '../util/arrayEncoding';
+import { createTableShareUrlFromTable } from '../util/arrayEncoding';
 
 export function useTableShare(tableId: number) {
   const { isOpen, openModal, closeModal, ModalWrapper } = useModal();
@@ -32,7 +32,7 @@ export function useTableShare(tableId: number) {
   // Process URL when data is successfully fetched
   useEffect(() => {
     if (data) {
-      setShareUrl(createTableShareUrl(baseUrl, data));
+      setShareUrl(createTableShareUrlFromTable(baseUrl, data));
     }
   }, [baseUrl, data]);
