+++ src/page/LandingPage/hooks/useLandingPageHandlers.ts
@@ -2,7 +2,7 @@ import { useNavigate } from 'react-router-dom';
 import { isLoggedIn } from '../../../util/accessToken';
 import { oAuthLogin } from '../../../util/googleAuth';
 import useLogout from '../../../hooks/mutations/useLogout';
-import { createTableShareUrl } from '../../../util/arrayEncoding';
+import { createTableShareUrlFromTable } from '../../../util/arrayEncoding';
 import { SAMPLE_TABLE_DATA } from '../../../constants/sample_table';
 import { useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -25,7 +25,7 @@ const useLandingPageHandlers = () => {
   // Declare functions that represent business logics
   const handleStartWithoutLogin = useCallback(() => {
     // window.location.href = LANDING_URLS.START_WITHOUT_LOGIN_URL;
-    window.location.href = createTableShareUrl(
+    window.location.href = createTableShareUrlFromTable(
       import.meta.env.VITE_SHARE_BASE_URL,
       SAMPLE_TABLE_DATA,
     );
